feat: Add liquidity to Uniswap
Add a new script, addLiquidity.ts, to add liquidity to Uniswap. This script allows users to add liquidity to the Uniswap pool by providing USDC and DAI tokens. The script uses the IERC20 and IUniswapV2Router contracts to interact with the tokens and the Uniswap router. Liquidity is added by calling the addLiquidity function with the specified amounts of USDC and DAI tokens, along with minimum amounts for each token. The script also logs the user's token balances before and after adding liquidity. This feature enhances the functionality of the project by enabling users to easily add liquidity to the Uniswap pool.